fix: persist scene data and add debugging overlay
